refactor : 중복되는 함수 제거 및 인터페이스로 구조 변경
